creating RightAlignedButtons component and refactoring CartButton into it
